v0.1.1-ALPHA
New feature: Keep Going. Similar to the Android version, the user can
select whether to start a new game or to keep playing this one after
he/she gets to 2048. This is what the online version lacks. v0.1-ALPHA
had 2048 lines of code, but now there are only 778 lines of code because
of the misusage of for loops and the generateTileObject function. Added
batch file which deletes the score. Also added Installer Builds for
downloading.